+++ infra/misc/svc_scripts/health_check.sh
@@ -1,14 +1,16 @@
+#!/bin/sh
 set -uf
 
 # Log to file
 exec >> "/var/log/health-checks.txt" 2>&1
 
-endpoint='127.0.0.1:8000/health/liveness'
+URL_PATH="$1"
+endpoint="127.0.0.1:8000${URL_PATH}"
 echo "Checking $endpoint"
-curl --fail -v "$endpoint"
+curl --fail --max-time 10 -v "$endpoint"
 EXIT_STATUS=$?
 if [ $EXIT_STATUS -ne 0 ]; then
-	echo "Health server liveness check failed"
+	echo "Health server essential check failed"
 	exit $EXIT_STATUS
 fi